just to save

+++ src/AAL/setcover.h
@@ -4,9 +4,11 @@ struct setcoverInstance {
   int* totalSet;
   unsigned int totalSetSize;
   int* subsets;
-  int* subsetSizes;
+  unsigned int* subsetSizes;
   unsigned int numberOfSubsets;
-  int* weights;
+  double* weights;
 };
 
-struct setcoverResult {};
+struct setcoverResult {
+
+};
